API/requirements.txt
asgiref==3.3.4
certifi==2021.10.8
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.4
django-countries==7.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
Pillow==8.2.0
PyJWT==2.1.0
pyparsing==3.0.6
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
uritemplate==4.1.1
urllib3==1.26.7
